Migrate the library to be Java 1.8 compatible unless you try to load classes on Folia
