+++ common/build.gradle
@@ -1,0 +1,7 @@
+group = 'com.tcoded.folialib'
+
+dependencies {
+    compileOnly 'org.spigotmc:spigot-api:1.8.8-R0.1-SNAPSHOT'
+
+    implementation(project(":platform:common"))
+}
