+++ platform/build.gradle
@@ -1,0 +1,1 @@
+group = 'com.tcoded.folialib'
