+++ platform/common/build.gradle
@@ -1,0 +1,5 @@
+group = 'com.tcoded.folialib.platform'
+
+dependencies {
+    compileOnly 'org.spigotmc:spigot-api:1.8.8-R0.1-SNAPSHOT'
+}
