+++ platform/common/src/main/java/com/tcoded/folialib/enums/EntityTaskResult.java
