+++ platform/common/src/main/java/com/tcoded/folialib/wrapper/task/WrappedTask.java
@@ -1,0 +1,14 @@
+package com.tcoded.folialib.wrapper.task;
+
+import org.bukkit.plugin.Plugin;
+
+@SuppressWarnings("unused")
+public interface WrappedTask {
+
+    void cancel();
+
+    boolean isCancelled();
+
+    Plugin getOwningPlugin();
+
+}
