+++ platform/paper/src/main/java/com/tcoded/folialib/impl/PaperImplementation.java
@@ -1,0 +1,13 @@
+package com.tcoded.folialib.impl;
+
+import com.tcoded.folialib.FoliaLib;
+
+@SuppressWarnings("unused")
+public class PaperImplementation extends SpigotImplementation {
+
+    public PaperImplementation(FoliaLib foliaLib) {
+        super(foliaLib);
+    }
+
+    // Don't need to override anything, since we're extending BukkitImplementation
+}
