+++ platform/spigot/build.gradle
@@ -1,0 +1,8 @@
+group = 'com.tcoded.folialib.platform'
+
+dependencies {
+    compileOnly 'org.spigotmc:spigot-api:1.8.8-R0.1-SNAPSHOT'
+
+    implementation(project(":common"))
+    implementation(project(":platform:common"))
+}
