--- src/main/java/com/tcoded/folialib/wrapper/WrappedTask.java
@@ -1,14 +1,0 @@
-package com.tcoded.folialib.wrapper;
-
-import org.bukkit.plugin.Plugin;
-
-@SuppressWarnings("unused")
-public interface WrappedTask {
-
-    void cancel();
-
-    boolean isCancelled();
-
-    Plugin getOwningPlugin();
-
-}
