--- src/main/java/com/tcoded/folialib/wrapper/task/WrappedBukkitTask.java
@@ -1,29 +1,0 @@
-package com.tcoded.folialib.wrapper.task;
-
-import com.tcoded.folialib.wrapper.WrappedTask;
-import org.bukkit.plugin.Plugin;
-import org.bukkit.scheduler.BukkitTask;
-
-public class WrappedBukkitTask implements WrappedTask {
-
-    private final BukkitTask task;
-
-    public WrappedBukkitTask(BukkitTask task) {
-        this.task = task;
-    }
-
-    @Override
-    public void cancel() {
-        this.task.cancel();
-    }
-
-    @Override
-    public boolean isCancelled() {
-        return this.task.isCancelled();
-    }
-
-    @Override
-    public Plugin getOwningPlugin() {
-        return this.task.getOwner();
-    }
-}
