--- src/main/java/com/tcoded/folialib/wrapper/task/WrappedFoliaTask.java
@@ -1,29 +1,0 @@
-package com.tcoded.folialib.wrapper.task;
-
-import com.tcoded.folialib.wrapper.WrappedTask;
-import io.papermc.paper.threadedregions.scheduler.ScheduledTask;
-import org.bukkit.plugin.Plugin;
-
-public class WrappedFoliaTask implements WrappedTask {
-
-    private final ScheduledTask task;
-
-    public WrappedFoliaTask(ScheduledTask task) {
-        this.task = task;
-    }
-
-    @Override
-    public void cancel() {
-        this.task.cancel();
-    }
-
-    @Override
-    public boolean isCancelled() {
-        return this.task.isCancelled();
-    }
-
-    @Override
-    public Plugin getOwningPlugin() {
-        return this.task.getOwningPlugin();
-    }
-}
